Fix bugs with ctors and fields that are checked but never created
It's possible for an app to have a type constructor that is
never actually used to create a value, yet is tested in a `case`.
This was generating invalid C.

The same thing is possible for records. My assumption was that all
fields must, at some point, be part of a record literal. Wrong!
The elm/time package defines an Era record but only as a type.
It never creates a value of that type, but it accesses its fields!
This happens in an unused branch of a used function!
It seems to be an unfinished feature for community to help finish.
Package does create an empty list of this record type!