Special-case code gen for saturated Basics function calls
Main motivation is `negate`, which cannot work without this!
In Basics.elm, negate x = -x
But its body `-x` generates `negate x`!
Cycle has to be broken using special case code gen
that replaces it with a kernel version.
In JS it would be replaced with JS - operator,
but in C we need to check Int/Float