Fix recursive functions in let expressions
These don't get generated as cycles, because JS doesn't really need to.
But it was breaking in C.
The self-reference is really a "free variable" from the outer scope, so
the Closure ends up containing a reference to its own address.
Need to allocate the Closure memory before inserting its free vars.
All free var insertion is now done this way just to reuse the same code,
even for anonymous functions that can't self-reference.